ai-testing-system/cost-controls/requirements.txt
# Cost Control System - Python Dependencies
# Perceptual Hashing Cache (Tier 2)

# Perceptual hashing
imagehash==4.3.1
Pillow==10.2.0

# Redis cache
redis==5.0.1

# PostgreSQL (for persistent storage)
psycopg2-binary==2.9.9
sqlalchemy==2.0.25

# AWS SDK (for S3 archival)
boto3==1.34.26

# Data validation
pydantic==2.5.3

# Utilities
python-dateutil==2.8.2

# Testing
pytest==7.4.4
pytest-mock==3.12.0

# Development
black==24.1.1
mypy==1.8.0

